Connected smartSort Function to Front End
Peer code with Halimo. Using jquery/ajax when an item is 'added' from the main page that item name (and associated user) is passed to a sorter route which calls the smartSort function. When a single match is found and returned, the item is automatically inserted into the database under that category and the count is updated on the page. When no matches are found that result is announced. When multiple matches are found, the matches are sent back to the front end as an array of objects.